backend/requirements.txt
# API framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.0
python-multipart==0.0.9

# Async task queue
arq==0.25.0
redis==5.0.4

# Ingestion
yt-dlp==2024.4.9
instaloader==4.10.3

# Audio transcription
openai-whisper==20231117
torch==2.3.0

# LLM (Ollama) + embeddings
ollama==0.3.3
voyageai==0.2.3

# MongoDB
motor==3.4.0
pymongo[srv]==4.7.0

# Media processing
ffmpeg-python==0.2.0
opencv-python-headless==4.9.0.80

# Utilities
httpx==0.27.0
tenacity==8.3.0
python-dotenv==1.0.1
